+++ cmd/bnsd/x/account/handler_test.go
@@ -32,6 +32,7 @@ func TestUseCases(t *testing.T) {
 		aliceCond   = weavetest.NewCondition()
 		bobCond     = weavetest.NewCondition()
 		charlieCond = weavetest.NewCondition()
+		brokerCond  = weavetest.NewCondition()
 
 		now = weave.UnixTime(1572247483)
 	)
@@ -52,6 +53,7 @@ func TestUseCases(t *testing.T) {
 							Admin:        aliceCond.Address(),
 							HasSuperuser: true,
 							AccountRenew: 1000,
+							Broker:       brokerCond.Address(),
 						},
 					},
 					BlockHeight: 1,
@@ -67,6 +69,7 @@ func TestUseCases(t *testing.T) {
 							Admin:        aliceCond.Address(),
 							HasSuperuser: true,
 							AccountRenew: 1000,
+							Broker:       brokerCond.Address(),
 						},
 					},
 					BlockHeight: 2,