+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `bnsapi`: move bnsapi to new repo
 - `bnscli`: fix boolean flag bug
 - `bugfix`: [issue#1165](https://github.com/iov-one/weave/issues/1165) 
+- `bnsd`: Expose msgfee bucket to the query router
 
 ## 1.0.0