+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `bnsapi`: move bnsapi to new repo
 - `bnscli`: fix boolean flag bug
 - `bug`: Refactor register domain message [issue#1165](https://github.com/iov-one/weave/issues/1165)
+- `bug`: Add broker to Account model and RegisterAccountMsg
 
 ## 1.0.0